refactor: add favorite tab
