+++ lib/pages/main_page.dart
@@ -1,17 +1,13 @@
-import 'package:dears/providers/portfolio_list_provider.dart';
+import 'package:dears/widgets/favorite_tab.dart';
 import 'package:dears/widgets/personal_tab.dart';
-import 'package:dears/widgets/portfolio_list_tile.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-class MainPage extends HookConsumerWidget {
+class MainPage extends HookWidget {
   const MainPage({super.key});
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final portfolioList = ref.watch(portfolioListProvider);
-
+  Widget build(BuildContext context) {
     final pageController = usePageController();
     final index = useState(0);
 
@@ -20,21 +16,11 @@ class MainPage extends HookConsumerWidget {
       body: PageView(
         controller: pageController,
         onPageChanged: (value) => index.value = value,
-        children: [
-          const Center(child: Text("홈")),
-          const Center(child: Text("검색")),
-          Center(
-            child: ListView(
-              children: List.generate(
-                portfolioList.length,
-                (index) {
-                  final portfolio = portfolioList[index];
-                  return PortfolioListTile(portfolio);
-                },
-              ),
-            ),
-          ),
-          const PersonalTab(),
+        children: const [
+          Center(child: Text("홈")),
+          Center(child: Text("검색")),
+          FavoriteTab(),
+          PersonalTab(),
         ],
       ),
       bottomNavigationBar: BottomNavigationBar(
